Torreta/software/requirements.txt
numpy
matplotlib
scipy
pyqtgraph

pylibiio
pyadi-iio
